Switch to a diagonal CKM matrix
Add a restriction card for the model with diagonal CKM matrix. It needs
to be put into $MADGRAPH/models/Massive_Higgs_UFO/.

This removes lots of diagrams that give virtually no contribution. We
also probably don't want to deal with events with decays W -> cb in the
context of tt reconstruction.